+++ src/main/java/step4/domain/Car.java
@@ -1,6 +1,8 @@
 package step4.domain;
 
 import step4.strategy.MoveStrategy;
+import step4.vo.Name;
+import step4.vo.Position;
 
 import java.util.Objects;
