+++ src/main/java/step4/domain/dto/GameHistoryDto.java
@@ -1,4 +1,4 @@
-package step4.service.dto;
+package step4.domain.dto;
 
 import step4.domain.GameHistory;
