+++ src/main/java/step4/service/Game.java
@@ -2,10 +2,10 @@ package step4.service;
 
 import step4.domain.Car;
 import step4.domain.GameHistory;
-import step4.service.dto.GameHistoryDto;
+import step4.domain.dto.GameHistoryDto;
 import step4.service.dto.GameInformation;
 import step4.strategy.MoveStrategy;
-import step4.domain.Names;
+import step4.vo.Names;
 import step4.domain.Cars;
 
 import java.util.List;
