+++ src/main/java/step4/service/dto/GameInformation.java
@@ -1,6 +1,6 @@
 package step4.service.dto;
 
-import step4.domain.Names;
+import step4.vo.Names;
 import step4.view.UserInput;
 
 public class GameInformation {
