+++ src/test/java/step4/NamesTest.java
@@ -4,8 +4,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import step4.domain.Name;
-import step4.domain.Names;
+import step4.vo.Name;
+import step4.vo.Names;
 
 import java.util.Arrays;
 import java.util.List;
