+++ src/test/java/step4/PositionTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import step4.domain.Position;
+import step4.vo.Position;
 
 import static org.assertj.core.api.Assertions.assertThat;
